requirements.txt
Django==1.6.5
django-widget-tweaks==1.3
psycopg2==2.5.3
wsgiref==0.1.2
south
django-haystack==2.2.0
pysolr==3.2.0
